% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thonk-package.R
\docType{package}
\name{thonk-package}
\alias{thonk-package}
\title{thonk: Explain and Address Error Messages}
\description{
Helps users understand and address error messages using large language models. With the tool enabled, errors raised to the user are accompanied by clickable links to "explain" or "fix" the issue. Explanations are printed to the console while fixes implement changes directly; in both cases, the model is supplied context about the files you're working in and the functions you're working with.
}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
